Move Components::Descriptions::* forms to Views::Controllers::Descriptions::*
Four sub-action forms move from app/components/descriptions/ to
their respective view-side homes under app/views/controllers/:

- AuthorRequestForm → Views::Controllers::Descriptions::AuthorRequests::Form
  (rendered by descriptions/author_requests#new)
- MergeForm → Views::Controllers::Descriptions::Merges::Form
  (rendered by {names,locations}/descriptions/merges#new)
- MoveForm → Views::Controllers::Descriptions::Moves::Form
  (rendered by {names,locations}/descriptions/moves#new)
- PermissionsForm → Views::Controllers::Descriptions::Permissions::Form
  (rendered by {names,locations}/descriptions/permissions#edit)

Merge/move/permissions are shared between Names::Descriptions and
Locations::Descriptions controllers (same sub-action surface, same
form), so they live under the top-level `descriptions/` namespace
rather than under either parent.

Explicit DOM ids ("merge_descriptions_form", "move_descriptions_form",
"description_permissions_form") are preserved via the existing
`super(id: "...")` overrides — integration tests and component tests
reference them by string, and the auto-derived ids
("description_merge_form", etc.) would be a separate change.

Test fixes carried along: dropped three `assert_match` / `scan` regex
assertions on HTML in the moved tests, replaced with the corresponding
`assert_html` / `assert_no_html` calls per the testing-rules ban on
regex-against-HTML.
