+++ app/views/controllers/locations/descriptions/moves/new.html.erb
@@ -5,4 +5,6 @@ add_context_nav(
 )
 %>
 
-<%= render(Components::Descriptions::MoveForm.new(@description, user: @user)) %>
+<%= render(Views::Controllers::Descriptions::Moves::Form.new(
+             @description, user: @user
+           )) %>
