+++ app/views/controllers/locations/descriptions/permissions/edit.html.erb
@@ -5,7 +5,7 @@ add_context_nav(
 )
 %>
 
-<%= render(Components::Descriptions::PermissionsForm.new(
+<%= render(Views::Controllers::Descriptions::Permissions::Form.new(
              description: @description,
              groups: @groups,
              data: @data
