+++ app/views/controllers/names/descriptions/moves/new.html.erb
@@ -3,4 +3,6 @@ add_page_title(:merge_descriptions_title.t(object: @description.format_name))
 add_context_nav(name_description_form_permissions_tabs(description: @description))
 %>
 
-<%= render(Components::Descriptions::MoveForm.new(@description, user: @user)) %>
+<%= render(Views::Controllers::Descriptions::Moves::Form.new(
+             @description, user: @user
+           )) %>
